requirements.txt
streamlit
boto3
ntplib
pillow
requests
streamlit-extras